[REA-181] feat(orders): add retrieve endpoint to CookerOrderView (#70)
* feat(orders): add retrieve endpoint to CookerOrderView

- Add RetrieveModelMixin to CookerOrderView
- Override retrieve() to use standardized success() response
- Add items_count field to CookerOrderGETSerializer for consistency with list serializer
- Consolidate retrieve tests into tests/cooker_app/orders/read/test_api.py
- Update orders_history expected_data fixtures with items_count field

* fix(customer_app): keep delivery distance in meters during order creation

* fix(customer_app): include delivery_fees as read-only in order creation response

* test(cooker_app): update sub_total to float in orders history tests

* feat(cooker_app): create CookerOrderHistorySerializer and use it for history retrieve

* fix(core_app): make delivery_fees mandatory on OrderModel

* fix(utils): return float from compute_order_items_total_amount to avoid type mismatches

* test(cooker_app): update order retrieve success test assertions

---------

+++ reats/utils/common.py
@@ -260,9 +260,9 @@ def create_stripe_customer(
             break
 
 
-def compute_order_items_total_amount(order: OrderModel) -> int:
-    dish_total = 0
-    drink_total = 0
+def compute_order_items_total_amount(order: OrderModel) -> float:
+    dish_total = 0.0
+    drink_total = 0.0
 
     for dish_item in order.dishes_items.all():  # type: ignore
         dish_total += dish_item.dish.price * dish_item.dish_quantity
